requirements.txt
python-dotenv==1.0.1
ipykernel==6.29.4
openpyxl==3.1.3
pandas==2.2.2